# fix(td-has-headers): greatly improve performance of td-has-headers rule