--- a/lib/checks/tables/td-has-header.js
+++ b/lib/checks/tables/td-has-header.js
@@ -1,6 +1,7 @@
-var tableUtils = axe.commons.table;
-var badCells = [];
-var cells = tableUtils.getAllCells(node);
+const tableUtils = axe.commons.table;
+const badCells = [];
+const cells = tableUtils.getAllCells(node);
+const tableGrid = tableUtils.toGrid(node);
 
 cells.forEach(cell => {
 	// For each non-empty data cell that doesn't have an aria label
@@ -10,7 +11,7 @@ cells.forEach(cell => {
 		!axe.commons.aria.label(cell)
 	) {
 		// Check if it has any headers
-		const hasHeaders = tableUtils.getHeaders(cell).some(header => {
+		const hasHeaders = tableUtils.getHeaders(cell, tableGrid).some(header => {
 			return header !== null && !!axe.commons.dom.hasContent(header);
 		});