--- a/lib/commons/table/get-cell-position.js
+++ b/lib/commons/table/get-cell-position.js
@@ -8,7 +8,7 @@
  * @param  {HTMLTableCellElement} cell The table cell of which to get the position
  * @return {Object} Object with `x` and `y` properties of the coordinates
  */
-table.getCellPosition = function(cell, tableGrid) {
+table.getCellPosition = axe.utils.memoize(function(cell, tableGrid) {
 	var rowIndex, index;
 	if (!tableGrid) {
 		tableGrid = table.toGrid(dom.findUp(cell, 'table'));
@@ -25,4 +25,4 @@ table.getCellPosition = function(cell, tableGrid) {
 			}
 		}
 	}
-};
+});