--- a/lib/commons/table/get-scope.js
+++ b/lib/commons/table/get-scope.js
@@ -29,7 +29,7 @@ table.getScope = function(cell) {
 		return false;
 	}
 	var tableGrid = table.toGrid(dom.findUp(cell, 'table'));
-	var pos = table.getCellPosition(cell);
+	var pos = table.getCellPosition(cell, tableGrid);
 
 	// The element is in a row with all th elements, that makes it a column header
 	var headerRow = tableGrid[pos.y].reduce((headerRow, cell) => {