--- a/lib/commons/table/to-grid.js
+++ b/lib/commons/table/to-grid.js
@@ -8,7 +8,8 @@
  * @param  {HTMLTableElement} node The table to convert
  * @return {Array<HTMLTableCellElement>} Array of HTMLTableCellElements
  */
-table.toGrid = function(node) {
+table.toGrid = axe.utils.memoize(function(node) {
+	console.log('cache miss!');
 	var table = [];
 	var rows = node.rows;
 	for (var i = 0, rowLength = rows.length; i < rowLength; i++) {
@@ -32,7 +33,7 @@ table.toGrid = function(node) {
 	}
 
 	return table;
-};
+});
 
 // This was the old name
 table.toArray = table.toGrid;