--- a/lib/commons/table/to-grid.js
+++ b/lib/commons/table/to-grid.js
@@ -9,7 +9,6 @@
  * @return {Array<HTMLTableCellElement>} Array of HTMLTableCellElements
  */
 table.toGrid = axe.utils.memoize(function(node) {
-	console.log('cache miss!');
 	var table = [];
 	var rows = node.rows;
 	for (var i = 0, rowLength = rows.length; i < rowLength; i++) {