--- a/lib/core/imports/index.js
+++ b/lib/core/imports/index.js
@@ -46,5 +46,6 @@ axe.imports = {
 	axios: require('axios'),
 	CssSelectorParser: require('css-selector-parser').CssSelectorParser,
 	doT: require('@deque/dot'),
-	emojiRegexText: require('emoji-regex')
+	emojiRegexText: require('emoji-regex'),
+	memoize: require('memoizee')
 };