--- a/lib/core/public/run-rules.js
+++ b/lib/core/public/run-rules.js
@@ -3,6 +3,7 @@
 
 // Clean up after resolve / reject
 function cleanup() {
+	axe._memoizedFns.forEach(fn => fn.clear());
 	axe._cache.clear();
 	axe._tree = undefined;
 	axe._selectorData = undefined;